Added scripts for parallelization

+++ run.sh
@@ -10,8 +10,11 @@
 # run.sh
 #   <full path to the program under test, e.g., $SCRIPT_DIR/../DafnyBench/DafnyBench/dataset/ground_truth/630-dafny_tmp_tmpz2kokaiq_Solution.dfy> 
 #   [--num_mutations <the number of mutations to apply to the input program, e.g., 1 (by default)>]
+#   [--run_dir <the directory where the script should be run, e.g., ./ (by deafult)>]
 #   [help]
 # ------------------------------------------------------------------------------ General utils
+
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" > /dev/null 2>&1 && pwd)"
 
 die() {
   echo "$@" >&2
@@ -23,8 +26,10 @@ die() {
 USAGE="Usage: ${BASH_SOURCE[0]}
    <full path to the program under test, e.g., $SCRIPT_DIR/../DafnyBench/DafnyBench/dataset/ground_truth/630-dafny_tmp_tmpz2kokaiq_Solution.dfy>
    [--num_mutations <the number of mutations to apply to the input program, e.g., 1 (by default)>]
+   [--run_dir <the directory where the script should be run, e.g., ./ (by deafult)>]
    [help]"
-if [ "$#" -ne "1" ] && [ "$#" -ne "3" ]; then
+
+if [ "$#" -ne "1" ] && [ "$#" -ne "3" ] && [ "$#" -ne "5" ]; then
   die "$USAGE"
 fi
 
@@ -33,28 +38,47 @@ if [ "$#" -eq "1" ] && [ "$1" = "--help" ]; then
     exit 0
 fi
 
-PROGRAM=$1
+PROGRAM=$1;
+PROGRAM="$(cd "$(dirname "$PROGRAM")" && pwd)/$(basename "$PROGRAM")"
+shift
 NUM_MUTS=1
-if [ "$#" -eq "3" ]; then
-    NUM_MUTS=$3
-fi
+RUN_DIR="./"
+while [[ "$1" = --* ]]; do
+  OPTION=$1; shift
+  case $OPTION in
+    (--num_mutations)
+      NUM_MUTS=$1;
+      shift;;
+    (--run_dir)
+      RUN_DIR=$1;
+      shift;;
+    (--help)
+      echo "$USAGE";
+      exit 0;;
+    (*)
+      die "$USAGE";;
+  esac
+done
+
+[ -d "$RUN_DIR" ] || die "[ERROR] $RUN_DIR does not exist!"
 
 # ------------------------------------------------------------------------------ Cleanup
 
 rm -rf targets.csv
-mkdir -p mutants
-mkdir -p mutants/alive
-mkdir -p mutants/timed-out
-mkdir -p mutants/killed
-mkdir -p mutants/invalid
+mkdir -p "$SCRIPT_DIR/original"
+mkdir -p "$SCRIPT_DIR/mutants"
+mkdir -p "$SCRIPT_DIR/mutants/alive"
+mkdir -p "$SCRIPT_DIR/mutants/timed-out"
+mkdir -p "$SCRIPT_DIR/mutants/killed"
+mkdir -p "$SCRIPT_DIR/mutants/invalid"
 
 # ------------------------------------------------------------------------------ MutDafny utils
 
 scan_program() {
     echo Scanning $PROGRAM for mutation targets
-    dotnet ./dafny/Binaries/Dafny.dll verify $PROGRAM \
-        --solver-path ./dafny/Binaries/z3 --allow-warnings \
-        --plugin ./mutdafny/bin/Debug/net8.0/mutdafny.dll,scan > /dev/null
+    dotnet "$SCRIPT_DIR/dafny/Binaries/Dafny.dll" verify $PROGRAM \
+        --solver-path "$SCRIPT_DIR/dafny/Binaries/z3" --allow-warnings \
+        --plugin "$SCRIPT_DIR/mutdafny/bin/Debug/net8.0/mutdafny.dll",scan > /dev/null
 }
 
 single_mutation() {
@@ -66,23 +90,23 @@ single_mutation() {
     if [[ -z $arg ]]; 
     then 
         echo Mutating position $pos: operator $op
-        output=$(dotnet ./dafny/Binaries/Dafny.dll verify $PROGRAM \
-            --solver-path ./dafny/Binaries/z3 --allow-warnings \
-            --plugin ./mutdafny/bin/Debug/net8.0/mutdafny.dll,"mut $pos $op" 2>/dev/null)
+        output=$(dotnet "$SCRIPT_DIR/dafny/Binaries/Dafny.dll" verify $PROGRAM \
+            --solver-path "$SCRIPT_DIR/dafny/Binaries/z3" --allow-warnings \
+            --plugin "$SCRIPT_DIR/mutdafny/bin/Debug/net8.0/mutdafny.dll","mut $pos $op" 2>/dev/null)
     else
         echo Mutating position $pos: operator $op, argument $arg
-        output=$(dotnet ./dafny/Binaries/Dafny.dll verify $PROGRAM \
-            --solver-path ./dafny/Binaries/z3 --allow-warnings \
-            --plugin ./mutdafny/bin/Debug/net8.0/mutdafny.dll,"mut $pos $op $arg" 2>/dev/null)
+        output=$(dotnet "$SCRIPT_DIR/dafny/Binaries/Dafny.dll" verify $PROGRAM \
+            --solver-path "$SCRIPT_DIR/dafny/Binaries/z3" --allow-warnings \
+            --plugin "$SCRIPT_DIR/mutdafny/bin/Debug/net8.0/mutdafny.dll","mut $pos $op $arg" 2>/dev/null)
     fi
     echo $output
 }
 
 multiple_mutation() {
     echo Applying $NUM_MUTS mutations to the program
-    output=$(dotnet ./dafny/Binaries/Dafny.dll verify $PROGRAM \
-        --solver-path ./dafny/Binaries/z3 --allow-warnings \
-        --plugin ./mutdafny/bin/Debug/net8.0/mutdafny.dll,"mut $NUM_MUTS" 2>/dev/null)
+    output=$(dotnet "$SCRIPT_DIR/dafny/Binaries/Dafny.dll" verify $PROGRAM \
+        --solver-path "$SCRIPT_DIR/dafny/Binaries/z3" --allow-warnings \
+        --plugin "$SCRIPT_DIR/mutdafny/bin/Debug/net8.0/mutdafny.dll","mut $NUM_MUTS" 2>/dev/null)
     echo $output
 }
 
@@ -99,7 +123,7 @@ process_output() {
 
         if [ -f *.dfy ]; then
             echo Error: mutant is invalid
-            mv *.dfy mutants/invalid
+            mv *.dfy "$SCRIPT_DIR/mutants/invalid"
         else
             echo Could not apply $NUM_MUTS mutations to the program
         fi
@@ -109,13 +133,13 @@ process_output() {
         output_dir=""
         if [[ -n $timed_out ]]; then
             echo Verification timed out
-            output_dir=mutants/timed-out
+            output_dir="$SCRIPT_DIR/mutants/timed-out"
         elif [[ -n $verified ]]; then 
             echo Verification succeeded: mutant is alive
-            output_dir=mutants/alive
+            output_dir="$SCRIPT_DIR/mutants/alive"
         else 
             echo Verification failed: mutant was killed
-            output_dir=mutants/killed
+            output_dir="$SCRIPT_DIR/mutants/killed"
         fi
 
         mv *.dfy $output_dir
@@ -126,10 +150,12 @@ process_output() {
 
 # ------------------------------------------------------------------------------ Main
 
+pushd . > /dev/null 2>&1
+cd "$RUN_DIR"
 scan_program
 
 IFS=','
-if [ "$#" -eq "1" ] || [ $NUM_MUTS -eq "1" ]; then
+if [ "$NUM_MUTS" -eq "1" ]; then
     while read pos op arg;
     do
 
@@ -177,3 +203,9 @@ else
 fi
 
 rm targets.csv
+cp ./original/*.dfy "$SCRIPT_DIR/original"
+popd > /dev/null 2>&1
+
+echo "[INFO] Job finished"
+echo "DONE!"
+exit 0
